requirement.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django
django-chartit==0.2.9
django-phonenumber-field
django-relativedelta
django-rest-framework==0.1.0
django-smsish==1.3.2
django-twilio
djangorestframework==3.12.1
idna==2.10
phonenumbers==8.12.10
plotly==4.11.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
twilio==6.45.4
urllib3==1.25.10
django-chartjs
django-crontab


